+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.14.1] - 2025-10-13
+- Fixed and issue with filtering resource exceptions by `--tags_filters`
+
 ## [5.14.0] - 2025-10-03
 - Added tenant settings key `SCAN_HIDDEN_REGIONS` that enables scanning of hidden regions
 - Implemented the division of findings by resource for pushing to Defect Dojo